requirements.txt
asgiref==3.2.3
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.9.0
django-extensions==2.2.9
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
numpy==1.18.3
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.3
Pillow==8.0.1
pretty-html-table==0.6.dev0
protobuf==3.12.2
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.3.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.2.0
